requirements.txt
Django==1.9.1
django-admin-tools==0.7.2
django-appconf==1.0.1
django-compressor==2.0
django-debug-toolbar==1.4
django-filebrowser==3.6.1
django-grappelli==2.7.3
django-modeltranslation==0.11
django-mptt==0.8.0
django-ratings==0.3.7
django-suit==0.2.16
django-suit-redactor==0.0.3
django-tastypie==0.13.1
django-treenav==1.0.0
easy-thumbnails==2.3
feedparser==5.2.1
FeinCMS==1.11.4
MySQL-python==1.2.5
Pillow==3.1.0
python-dateutil==2.4.2
python-mimeparse==1.5
pytz==2015.7
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sqlparse==0.1.18
